pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "solar-grader"
version = "0.1.0"
description = "Automatic Grading System for Coding Assignments using Upstage Solar & Document Parse"
readme = "README.md"
authors = [
  { name = "Your Name", email = "your.email@example.com" },
]
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "streamlit>=1.30.0",
    "pandas>=2.0.0",
    "openai>=1.52.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "PyPDF2>=3.0.0",
    "loguru>=0.7.0",
]

[project.urls]
"Homepage" = "https://github.com/your-repo/solar-grader"

[project.scripts]
solar-grader = "solar_grader.__main__:main"
